[driver/opc] refactor reader.cpp into source and task
